- refactor schedule after and before reboot   - have to change one more thing in regards to when we show the forground notificaiton
